requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
Werkzeug==3.1.3
WTForms==3.2.1
cryptography==44.0.2
python-dotenv==1.0.1
langchain==0.3.21
langchain-groq==0.3.1
langchain-community==0.3.20
